+++ scripts/markdown.js
@@ -1,4 +1,4 @@
-import { EditorState } from "@codemirror/state";
+import { Compartment, EditorState } from "@codemirror/state";
 import { EditorView } from "@codemirror/view";
 import { markdown } from "@codemirror/lang-markdown";
 import {
@@ -21,6 +21,7 @@ import {
 } from "@codemirror/autocomplete";
 import { languages } from "@codemirror/language-data";
 
+import theme from "./codemirror_theme.js";
 import * as ui from "./markdown_ui.js";
 
 const makeBold = ui.toggleTag("**", "**");
@@ -51,9 +52,12 @@ const markdownBinding = [
 ];
 
 export function init(parent, textarea, pasteLink = createLink) {
+  const themeConfig = new Compartment();
+  const initTheme = document.documentElement.dataset.theme == "dark" ? theme.dark : theme.light;
   const state = EditorState.create({
     doc: textarea.value,
     extensions: [
+      themeConfig.of(initTheme),
       highlightSpecialChars(),
       history(),
       dropCursor(),
@@ -107,6 +111,13 @@ export function init(parent, textarea, pasteLink = createLink) {
 
   textarea.form.addEventListener("submit", () => {
     textarea.value = editor.state.doc.toString();
+  });
+
+  addEventListener("themeChange", (event) => {
+    const themeName = event.detail.theme === "dark" ? "dark" : "light";
+    editor.dispatch({
+      effects: themeConfig.reconfigure(theme[themeName]),
+    });
   });
 
   return {