+++ static/libs/markdown.js
@@ -1,1 +1,1 @@
-export{w as init}from"./markdown-C9iaJ-TC.js";
+export{w as init}from"./markdown-nGedf7mj.js";